+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This changelog consists of the bug & security fixes and new features being inclu
 
 * #10735 [fixed] - Fixed the issue with the multiselect filter that was preventing products from being correctly filtered on the category page.
 
+* #10718 [fixed] - Fixed the issue that allowed an admin to delete their own account.
+
 ## **v2.3.3 (22nd of May 2025)** - *Release*
 
 * Resolved an issue with the category filter functionality.